refactor: assets move to external
