+++ assets/resourceId.go
