+++ assets/resourceId_test.go
