+++ deployment/managementgroup_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/Azure/alzlib"
-	"github.com/Azure/alzlib/internal/assets"
+	"github.com/Azure/alzlib/assets"
 	"github.com/Azure/alzlib/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/authorization/armauthorization"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/resources/armpolicy"
@@ -396,7 +396,7 @@ func TestModifyPolicySetDefinitions(t *testing.T) {
 		"pd1": "mg1",
 	}
 	err := updatePolicySetDefinitions(alzmg, pd2mg)
-	assert.NoError(t, err)
+	require.NoError(t, err)
 	expected := fmt.Sprintf(PolicySetDefinitionIdFmt, "mg1", "psd1")
 	assert.Equal(t, expected, *alzmg.policySetDefinitions["psd1"].ID)
 	expected = fmt.Sprintf(PolicyDefinitionIdFmt, "mg1", "pd1")
